backend/requirements.txt
# --- Dependencias del Backend ---

# Framework principal
Django

# Para crear la API REST
djangorestframework

# Conector para la base de datos PostgreSQL
psycopg2-binary

# Para la documentación automática de la API (Swagger UI)
drf-spectacular

# Para permitir que el frontend se comunique con el backend (CORS)
django-cors-headers

# Para parsear la URL de la base de datos desde variables de entorno
dj-database-url
dotenv